refactor: 패키지 구조 변경
- theme stat 이동 취소
