+++ src/main/java/com/ddobang/backend/domain/theme/repository/ThemeRepositoryImpl.java
@@ -9,10 +9,10 @@ import com.ddobang.backend.domain.store.entity.QStore;
 import com.ddobang.backend.domain.theme.dto.request.ThemeFilterRequest;
 import com.ddobang.backend.domain.theme.dto.response.SimpleThemeResponse;
 import com.ddobang.backend.domain.theme.entity.QTheme;
+import com.ddobang.backend.domain.theme.entity.QThemeStat;
 import com.ddobang.backend.domain.theme.entity.QThemeTagMapping;
 import com.ddobang.backend.domain.theme.entity.Theme;
-import com.ddobang.backend.domain.theme.stat.entity.QThemeStat;
-import com.ddobang.backend.domain.theme.stat.entity.ThemeStat;
+import com.ddobang.backend.domain.theme.entity.ThemeStat;
 import com.ddobang.backend.domain.theme.tag.entity.QThemeTag;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Projections;
