+++ src/test/java/com/ddobang/backend/domain/theme/controller/ThemeControllerTest.java
@@ -21,9 +21,9 @@ import com.ddobang.backend.domain.region.entity.Region;
 import com.ddobang.backend.domain.store.entity.Store;
 import com.ddobang.backend.domain.theme.dto.request.ThemeFilterRequest;
 import com.ddobang.backend.domain.theme.entity.Theme;
-import com.ddobang.backend.domain.theme.stat.entity.ThemeStat;
+import com.ddobang.backend.domain.theme.entity.ThemeStat;
 import com.ddobang.backend.domain.theme.exception.ThemeErrorCode;
-import com.ddobang.backend.domain.theme.stat.repository.ThemeStatRepository;
+import com.ddobang.backend.domain.theme.repository.ThemeStatRepository;
 import com.ddobang.backend.domain.theme.tag.entity.ThemeTag;
 import com.ddobang.backend.global.initdata.BaseInitData;
 import com.fasterxml.jackson.databind.ObjectMapper;
