+++ src/test/java/com/ddobang/backend/domain/theme/repository/ThemeRepositoryTest.java
@@ -20,8 +20,7 @@ import com.ddobang.backend.domain.store.repository.StoreRepository;
 import com.ddobang.backend.domain.theme.dto.request.ThemeFilterRequest;
 import com.ddobang.backend.domain.theme.dto.response.SimpleThemeResponse;
 import com.ddobang.backend.domain.theme.entity.Theme;
-import com.ddobang.backend.domain.theme.stat.entity.ThemeStat;
-import com.ddobang.backend.domain.theme.stat.repository.ThemeStatRepository;
+import com.ddobang.backend.domain.theme.entity.ThemeStat;
 import com.ddobang.backend.domain.theme.tag.entity.ThemeTag;
 import com.ddobang.backend.domain.theme.tag.repository.ThemeTagRepository;
 import com.ddobang.backend.global.config.JpaAuditingConfig;
